+++ src/ts/magma/api/collect/List.ts
@@ -4,6 +4,7 @@
 	Collector: magma.api.collect, 
 	EmptyHead: magma.api.collect, 
 	Head: magma.api.collect, 
+	JVMList: magma.api.collect, 
 	List: magma.api.collect, 
 	ListCollector: magma.api.collect, 
 	Lists: magma.api.collect, 
@@ -44,4 +45,5 @@ export interface List<T> {
 	isEmpty(): boolean;
 	equalsTo(other: List<T>): boolean;
 	removeValue(element: T): List<T>;
+	removeLast(): Option<List<T>>;
 }
