+++ src/ts/magma/api/collect/ListCollector.ts
@@ -4,6 +4,7 @@
 	Collector: magma.api.collect, 
 	EmptyHead: magma.api.collect, 
 	Head: magma.api.collect, 
+	JVMList: magma.api.collect, 
 	List: magma.api.collect, 
 	ListCollector: magma.api.collect, 
 	Lists: magma.api.collect, 
@@ -25,8 +26,8 @@
 	Main: magma.app, 
 	Files: magma.jvm.io
 ]*/
-import { List } from "../../../magma/api/collect/List";
 import { Collector } from "../../../magma/api/collect/Collector";
+import { List } from "../../../magma/api/collect/List";
 import { Lists } from "../../../magma/api/collect/Lists";
 export class ListCollector<T> implements Collector<T, List<T>> {
 	createInitial(): List<T> {
