+++ src/ts/magma/api/collect/Lists.ts
@@ -4,6 +4,7 @@
 	Collector: magma.api.collect, 
 	EmptyHead: magma.api.collect, 
 	Head: magma.api.collect, 
+	JVMList: magma.api.collect, 
 	List: magma.api.collect, 
 	ListCollector: magma.api.collect, 
 	Lists: magma.api.collect, 
@@ -27,12 +28,6 @@
 ]*/
 import { Actual } from "../../../magma/annotate/Actual";
 import { Namespace } from "../../../magma/annotate/Namespace";
-import { Tuple2 } from "../../../magma/api/Tuple2";
-import { Tuple2Impl } from "../../../magma/api/Tuple2Impl";
-import { HeadedQuery } from "../../../magma/api/collect/query/HeadedQuery";
-import { Query } from "../../../magma/api/collect/query/Query";
-import { Option } from "../../../magma/api/option/Option";
-import { Some } from "../../../magma/api/option/Some";
 import { ArrayList } from "../../../java/util/ArrayList";
 import { Arrays } from "../../../java/util/Arrays";
 import { List } from "../../../magma/api/collect/List";
